Realizando criação do campo Categoria na adição de um item

+++ EstoqueTIRDEB/Services/ItensService.cs
@@ -1,9 +1,7 @@
 ﻿using EstoqueTIRDEB.Data;
 using EstoqueTIRDEB.Models;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace EstoqueTIRDEB.Services
 {
@@ -15,7 +13,6 @@ namespace EstoqueTIRDEB.Services
         {
             _context = context;
         }
-
 
         public List<Itens> FindAll()
         {
@@ -30,14 +27,37 @@ namespace EstoqueTIRDEB.Services
         public void Remove(int id)
         {
             var obj = _context.Itens.Find(id);
-            _context.Itens.Remove(obj);
-            _context.SaveChanges();
+            if (obj != null)
+            {
+                _context.Itens.Remove(obj);
+                _context.SaveChanges();
+            }
         }
 
         public void Insert(Itens obj)
         {
             _context.Add(obj);
             _context.SaveChanges();
+        }
+
+        public List<Itens> ListarPorCategoria(int categoriaId)
+        {
+            return _context.Itens.Where(i => i.CategoriaId == categoriaId).ToList();
+        }
+
+        public void AdicionarItemAoEstoque(Itens item, int categoriaId)
+        {
+            var categoria = _context.Categoria.Find(categoriaId);
+            if (categoria != null)
+            {
+                item.CategoriaId = categoriaId;
+                _context.Itens.Add(item);
+                _context.SaveChanges();
+            }
+            else
+            {
+                // Tratar erro de categoria não encontrada
+            }
         }
     }
 }
